backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
python-dotenv==1.0.0
PyJWT>=2.8.0              # JWT authentication
# psycopg2-binary==2.9.9  # Uncomment for PostgreSQL support

# LLM Provider SDKs
anthropic>=0.39.0
openai>=1.54.0
requests>=2.32.0

# Document Parsing Libraries
pdfplumber>=0.11.0        # PDF text extraction (recommended)
PyPDF2>=3.0.0             # Alternative PDF library
python-docx>=1.1.0        # Word document support
pytesseract>=0.3.10       # OCR for images
Pillow>=10.0.0            # Image processing for OCR
